Add bookings

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -6,11 +6,11 @@ import java.util.List;
 public interface UserService {
     UserDto create(UserDto userDto);
 
-    UserDto update(Integer id, UserDto userDto);
+    UserDto update(Long id, UserDto userDto);
 
-    UserDto findById(Integer id);
+    UserDto findById(Long id);
 
     List<UserDto> findAll();
 
-    void delete(Integer id);
+    void delete(Long id);
 }
